+++ chess/king.py
@@ -6,12 +6,7 @@ class King(Piece):
    black_str = "♚"
    
    def get_valid_moves(self, from_row, from_col, to_row, to_col):
-      return self.is_valid_king_move(from_row, from_col, to_row, to_col)
-
-   def is_valid_king_move(self, from_row, from_col, to_row, to_col):
-     # Movimientos ortogonales y diaagonales 
-     possible_move= self.calculate_king_moves(from_row, from_col)
-     return (to_row, to_col)in possible_move
+      return (to_row, to_col) in self.calculate_king_moves(from_row, from_col)
    
    def calculate_king_moves(self, from_row, from_col):
         # Solo permite un paso en cada dirección