+++ tests/tests_cli.py
@@ -1,6 +1,6 @@
 import unittest
 from unittest.mock import patch
-from chess.chess import Chess
+from chess.ajedrez import Chess
 from chess.cli import play      #patch sobreescribe el comportamiento de algo, en este caso sobreescribe el print, el imput no se ejecuta
 
 
@@ -21,7 +21,7 @@ class TestCli(unittest.TestCase):
         chess = Chess()
         play(chess)
         self.assertEqual(mock_input.call_count, 4)
-        self.assertEqual(mock_print.call_count, 3)
+        self.assertEqual(mock_print.call_count, 2)
         self.assertEqual(mock_chess_move.call_count, 1)
         
     @patch( # este patch controla lo que hace el input